Make Graph and Workspace related debugging tools available in `but-testing`
