+++ crates/but-graph/tests/graph/init/mod.rs
@@ -1,6 +1,5 @@
-use crate::graph_tree;
 use but_graph::Graph;
-use but_testsupport::visualize_commit_graph_all;
+use but_testsupport::{graph_tree, graph_workspace, visualize_commit_graph_all};
 
 #[test]
 fn unborn() -> anyhow::Result<()> {
@@ -710,7 +709,6 @@ mod with_workspace;
 
 mod utils;
 use crate::init::utils::standard_options_with_extra_target;
-use crate::vis::utils::graph_workspace;
 pub use utils::{
     StackState, add_stack_with_segments, add_workspace, id_at, id_by_rev,
     read_only_in_memory_scenario, standard_options,
