+++ crates/but-graph/tests/graph/main.rs
@@ -1,5 +1,4 @@
 mod init;
 mod vis;
-pub use vis::utils::graph_tree;
 
 mod ref_metadata_legacy;
