+++ crates/but-testsupport/Cargo.toml
@@ -13,5 +13,6 @@ gix.workspace = true
 anyhow.workspace = true
 termtree = "0.5.1"
 gix-testtools.workspace = true
+but-graph.workspace = true
 
 [dev-dependencies]
